# Issue 4196: Increase batch size in the client side for throughput-oriented workloads