**Change log description**  
In crease batch size on the client side to provide better performance on throughput-oriented workloads.

**Purpose of the change**  
Partially fixes #4196.

**What the code does** 
The current batching algorithm in the client defines the size of the batches based on:
- The number of appends (and their size) waiting in the queue;
- The maximum defined batch size is full.

This algorithm attempts to handle both latency-oriented (1st condition is met) and throughput-oriented workloads (2nd condition is met).

However, the current max size of the batch is rather small for throughput-oriented workloads: 32KB. The consequence of this is that, when a writer writes a moderate/high amount of events (e.g., 5K e/s) of a moderate size (e.g., 10KB), we observe a severe increase in latency of write operations. The root cause is that the batching algorithm is all the time using 32KB batches, as can be observed in this plot:

![batch_size_varying_rate](https://user-images.githubusercontent.com/717112/65898696-6e8e1580-e3b2-11e9-9f80-b2137a32dda3.jpg)

However, it looks like the client is hitting a ceiling with the 32KB max batch size that yields a to increased queuing, and therefore more latency from a client perspective.

In this PR, we increase the max batch size to better accommodate throughput-oriented workloads, while keeping the behavior of the client as it is for latency oriented workloads.

**How to verify it**  

The change proposed in this PR has important improvements in terms of throughput and latency for throughput-oriented workloads:

_Unthrottled workload:_
```
Original:
INFO io.pravega.perf.PerfStats - 120347 records Writing, 4098.454 records/sec, 10000 bytes record size, 39.09 MiB/sec, 187.0 ms avg latency, 303.0 ms max latency, 185 ms 50th, 205 ms 75th, 218 ms 95th, 268 ms 99th, 278 ms 99.9th, 279 ms 99.99th.
This PR:
INFO io.pravega.perf.PerfStats - 283020 records Writing, 9657.738 records/sec, 10000 bytes record size, 92.10 MiB/sec, 72.4 ms avg latency, 342.0 ms max latency, 69 ms 50th, 81 ms 75th, 107 ms 95th, 139 ms 99th, 303 ms 99.9th, 309 ms 99.99th.
```
For latency-oriented workloads, the behavior of the client remains the same:

_Throttled workload:_
```
Original:
2019-09-30 13:51:40:252 +0000 [ForkJoinPool-1-worker-1] INFO io.pravega.perf.PerfStats - 29199 records Writing, 999.589 records/sec, 10000 bytes record size, 9.53 MiB/sec, 10.2 ms avg latency, 71.0 ms max latency, 10 ms 50th, 12 ms 75th, 14 ms 95th, 15 ms 99th, 25 ms 99.9th, 33 ms 99.99th.
This PR:
2019-09-30 13:50:42:810 +0000 [ForkJoinPool-1-worker-1] INFO io.pravega.perf.PerfStats - 29227 records Writing, 999.624 records/sec, 10000 bytes record size, 9.53 MiB/sec, 9.7 ms avg latency, 72.0 ms max latency, 9 ms 50th, 11 ms 75th, 13 ms 95th, 20 ms 99th, 44 ms 99.9th, 49 ms 99.99th.
```
No logic changes, all tests should be passing.