@RaulGracia 2019-10-01
@tkaitchuck @shrids in the following I post the results of some experiments with Pravega Benchmark, using both `master` version of  the Pravega client and the version of this PR in an on-premise PKS environment. The objective is to show, for a different and more limited environment compared to the previous experiment, that when the client batch size is the limiting factor, this PR provides significant benefits, whereas for the other cases the behavior remains the same: 

``` 
1 Writer - 1 Segment - Unthrottled client
Master (5 minutes):
2019-10-01 13:41:31:320 +0000 [ForkJoinPool-1-worker-1] INFO io.pravega.perf.PerfStats - 411372 records Writing, 1373.066 records/sec, 10000 bytes record size, 13.09 MiB/sec, 524.2 ms avg latency, 4211.0 ms max latency, 480 ms 50th, 520 ms 75th, 791 ms 95th, 1882 ms 99th, 3076 ms 99.9th, 3329 ms 99.99th.

This PR (5 minutes):
2019-10-01 13:52:21:535 +0000 [ForkJoinPool-1-worker-1] INFO io.pravega.perf.PerfStats - 595889 records Writing, 1999.674 records/sec, 10000 bytes record size, 19.07 MiB/sec, 85.9 ms avg latency, 2138.0 ms max latency, 32 ms 50th, 77 ms 75th, 293 ms 95th, 674 ms 99th, 1120 ms 99.9th, 1218 ms 99.99th
```

Despite the important limitations of Tier 1 performance in this environment, we observe that a larger batch size can improve throughput oriented workloads as well. The improvement in throughput is over 45%, whereas the latency reduction is -85% and -64% for percentiles 75th and 99th, respectively.

```
1 Writer - 10 Segments - Throttled to 2K e/s
Master (5 minutes):
2019-10-01 14:15:31:015 +0000 [ForkJoinPool-1-worker-1] INFO io.pravega.perf.PerfStats - 597434 records Writing, 1999.197 records/sec, 10000 bytes record size, 19.07 MiB/sec, 63.6 ms avg latency, 703.0 ms max latency, 51 ms 50th, 66 ms 75th, 147 ms 95th, 318 ms 99th, 450 ms 99.9th, 489 ms 99.99th.

This PR (5 minutes):
2019-10-01 14:20:53:120 +0000 [ForkJoinPool-1-worker-1] INFO io.pravega.perf.PerfStats - 597850 records Writing, 1999.191 records/sec, 10000 bytes record size, 19.07 MiB/sec, 65.9 ms avg latency, 661.0 ms max latency, 54 ms 50th, 73 ms 75th, 133 ms 95th, 311 ms 99th, 505 ms 99.9th, 548 ms 99.99th.
```
```
1 Writer - 10 Segments - Unthrottled client
Master (5 minutes):
2019-10-01 14:33:04:151 +0000 [ForkJoinPool-1-worker-1] INFO io.pravega.perf.PerfStats - 1248941 records Writing, 4149.292 records/sec, 10000 bytes record size, 39.57 MiB/sec, 796.4 ms avg latency, 7416.0 ms max latency, 543 ms 50th, 1094 ms 75th, 2354 ms 95th, 3489 ms 99th, 4400 ms 99.9th, 4584 ms 99.99th.

This PR (5 minutes):
2019-10-01 14:27:13:690 +0000 [ForkJoinPool-1-worker-1] INFO io.pravega.perf.PerfStats - 1237039 records Writing, 4129.768 records/sec, 10000 bytes record size, 39.38 MiB/sec, 729.0 ms avg latency, 5436.0 ms max latency, 517 ms 50th, 967 ms 75th, 2077 ms 95th, 2990 ms 99th, 3815 ms 99.9th, 4073 ms 99.99th.
```
In these 2 experiments, we observe that when we throttle the client way below the saturation point, or the client can saturate Tier 1 (increasing segments, and therefore parallelism for Tier 1 writes), the performance of the client remains the same.