@RaulGracia 2019-10-02
@tkaitchuck we have already observed throughput and latency improvements in both PKS (on-premise) and AWS. System tests are also green. Please, consider approving and merging this PR.